增加了基于角色的访问控制

+++ src/main/java/com/example/testonline/controller/ChoiceController.java
@@ -2,8 +2,10 @@ package com.example.testonline.controller;
 
 
 import com.example.testonline.pojo.Choicequsetion;
+import com.example.testonline.result.CodeEnum;
 import com.example.testonline.result.Result;
 import com.example.testonline.result.ResultUtil;
+import com.example.testonline.result.TosException;
 import com.example.testonline.service.ChoiceQuestionService;
 import com.example.testonline.service.ReadService;
 import org.springframework.stereotype.Controller;
@@ -13,6 +15,8 @@ import org.springframework.web.multipart.MultipartFile;
 import javax.annotation.Resource;
 import java.util.List;
 import java.util.Map;
+
+import static com.example.testonline.shiro.ShiroUtil.hasRole;
 
 @Controller
 @RequestMapping(value = "/templates")
@@ -27,6 +31,7 @@ public class ChoiceController {
     @RequestMapping(value = "/insertChoicequsetion",method = RequestMethod.POST)
     @ResponseBody
     public Result insertChoice(@RequestBody Choicequsetion choicequsetion) {
+        if (!hasRole("admin","teacher")) throw new TosException(CodeEnum.JurisdictionException);
         choiceQuestionService.insert(choicequsetion);
         return ResultUtil.sussess();
     }
@@ -34,6 +39,8 @@ public class ChoiceController {
     @RequestMapping(value = "/insertChoiceByExcel", method = RequestMethod.POST)
     @ResponseBody
     public Result insertChoiceByExcel(@RequestParam("file") MultipartFile file) {
+
+        if (!hasRole("admin","teacher")) throw new TosException(CodeEnum.JurisdictionException);
         List<Object> datas = readService.excelReadService(file, Choicequsetion.class);
         Choicequsetion choicequsetion;
         for (Object data:
@@ -62,6 +69,7 @@ public class ChoiceController {
     @RequestMapping(value = "/updateChoicequsetion",method = RequestMethod.POST)
     @ResponseBody
     public Result updateChoice(@RequestBody Choicequsetion choicequsetion) {
+        if (!hasRole("admin","teacher")) throw new TosException(CodeEnum.JurisdictionException);
         choiceQuestionService.insert(choicequsetion);
         return ResultUtil.sussess();
     }
@@ -69,6 +77,7 @@ public class ChoiceController {
     @RequestMapping(value = "/deleteChoicequestion",method = RequestMethod.POST)
     @ResponseBody
     public Result deleteChoice(@RequestBody Map<String, String> person) {
+        if (!hasRole("admin","teacher")) throw new TosException(CodeEnum.JurisdictionException);
         String id = person.get("id");
         choiceQuestionService.delete(Integer.valueOf(id));
         return ResultUtil.sussess();
